pyproject.toml
[tool.poetry]
name = "singlepageblog"
version = "0.1.0"
description = ""
authors = ["Neeraj319 <itskop520@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.1"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
